tests/wpt/mozilla/tests/mozilla/characterdata.html

<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>This is the character data</p>
  </body>
  <script>
  test(function() {
    var a = document.getElementsByTagName('p')[0].childNodes[0];
    assert_equals(a.data, "This is the character data");
    // append test utf8
    a.appendData(", append more 資料，測試資料");
    assert_equals(a.data, "This is the character data, append more 資料，測試資料");

    // length test utf8
    assert_equals(a.length, 47);

    // insert test non-utf8
    a.insertData(26, " test");
    assert_equals(a.data, "This is the character data test, append more 資料，測試資料");
    // insert test utf8
    a.insertData(48, "更多");
    assert_equals(a.data, "This is the character data test, append more 資料，更多測試資料");

    // delete test non-utf8
    a.deleteData(40, 5);
    assert_equals(a.data, "This is the character data test, append 資料，更多測試資料");
    // delete test utf8
    a.deleteData(45, 2);
    assert_equals(a.data, "This is the character data test, append 資料，更多資料");

    // replace test non-utf8
    a.replaceData(33, 6, "other");
    assert_equals(a.data, "This is the character data test, other 資料，更多資料");
    // replace test non-utf8
    a.replaceData(44, 2, "文字");
    assert_equals(a.data, "This is the character data test, other 資料，更多文字");

    // substring test non-utf8
    assert_equals(a.substringData(12, 4), "char");
    // substring test utf8
    assert_equals(a.substringData(39, 2), "資料");
  });
  </script>
</html>
